Create Explorer.SmartContract.Reader
This new module is the main entry point for getting results of a smart
contract function. It accepts a smart contract hash and a list of
functions names and its arguments (when applicable) in order to execute
those functions and return their results.

Before actually making the call to the blockchain, the function name is
encrypted and the arguments are encoded. The :abi lib deals with this,
encrypting the name of the function using the Keccak256 algorithm and
encoding the arguments based on their type.

The call to the blockchain is made by the EthereumJSONRPC app in the
umbrella.

The result is then decoded - as it comes in the same type of encoding
that has been sent - by the :abi lib as well, and returned. The return
is a list of results of calling the functions in the blockchain.